requirements.txt
fastapi[standard]
-q -U google-generativeai
streamlite
pydantic
uvicorn
